# services/playbookmanagement/src/main/java/nl/utwente/sosoc/playbookmanagement/api/PlaybooksApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.11.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package nl.utwente.sosoc.playbookmanagement.api;

import nl.utwente.sosoc.playbookmanagement.model.Playbook;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-03-19T22:15:24.388401400+01:00[Europe/Amsterdam]", comments = "Generator version: 7.11.0")
@Validated
@Tag(name = "playbooks", description = "the playbooks API")
public interface PlaybooksApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * DELETE /playbooks/{id} : Delete an existing playbook
     *
     * @param id Playbook ID (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "deletePlaybook",
        summary = "Delete an existing playbook",
        tags = { "playbooks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/playbooks/{id}"
    )
    
    default ResponseEntity<Void> deletePlaybook(
        @Parameter(name = "id", description = "Playbook ID", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /playbooks/{id} : Get an existing playbook
     *
     * @param id Playbook ID (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getPlaybook",
        summary = "Get an existing playbook",
        tags = { "playbooks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Playbook.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/playbooks/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Playbook> getPlaybook(
        @Parameter(name = "id", description = "Playbook ID", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"firstStep\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"name\" : \"phishing email\", \"description\" : \"quarantines endpoint\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"trigger\" : { \"match\" : \"Threat.Code == 'phishing'\" }, \"steps\" : [ { \"next\" : [ { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" }, { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" } ], \"name\" : \"await analyst confirmation\", \"action\" : \"action\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"type\" : \"human\" }, { \"next\" : [ { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" }, { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" } ], \"name\" : \"await analyst confirmation\", \"action\" : \"action\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"type\" : \"human\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /playbooks : Get all playbooks
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getPlaybooks",
        summary = "Get all playbooks",
        tags = { "playbooks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Playbook.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/playbooks",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Playbook>> getPlaybooks(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"firstStep\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"name\" : \"phishing email\", \"description\" : \"quarantines endpoint\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"trigger\" : { \"match\" : \"Threat.Code == 'phishing'\" }, \"steps\" : [ { \"next\" : [ { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" }, { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" } ], \"name\" : \"await analyst confirmation\", \"action\" : \"action\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"type\" : \"human\" }, { \"next\" : [ { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" }, { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" } ], \"name\" : \"await analyst confirmation\", \"action\" : \"action\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"type\" : \"human\" } ] }, { \"firstStep\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"name\" : \"phishing email\", \"description\" : \"quarantines endpoint\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"trigger\" : { \"match\" : \"Threat.Code == 'phishing'\" }, \"steps\" : [ { \"next\" : [ { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" }, { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" } ], \"name\" : \"await analyst confirmation\", \"action\" : \"action\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"type\" : \"human\" }, { \"next\" : [ { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" }, { \"next\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"condition\" : \"condition\" } ], \"name\" : \"await analyst confirmation\", \"action\" : \"action\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"type\" : \"human\" } ] } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /playbooks : Create a new playbook
     *
     * @param playbook  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "postPlaybook",
        summary = "Create a new playbook",
        tags = { "playbooks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/playbooks",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> postPlaybook(
        @Parameter(name = "Playbook", description = "", required = true) @Valid @RequestBody Playbook playbook
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /playbooks/{id} : Update an existing playbook
     *
     * @param id Playbook ID (required)
     * @param playbook  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "putPlaybook",
        summary = "Update an existing playbook",
        tags = { "playbooks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/playbooks/{id}",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> putPlaybook(
        @Parameter(name = "id", description = "Playbook ID", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "Playbook", description = "", required = true) @Valid @RequestBody Playbook playbook
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}