# services/threatintelligence/src/main/java/nl/utwente/sosoc/threatintelligence/api/AlarmsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.11.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package nl.utwente.sosoc.threatintelligence.api;

import nl.utwente.sosoc.threatintelligence.model.Alarm;
import nl.utwente.sosoc.threatintelligence.model.IOC;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-03-19T22:15:26.710031900+01:00[Europe/Amsterdam]", comments = "Generator version: 7.11.0")
@Validated
@Tag(name = "alarms", description = "the alarms API")
public interface AlarmsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /alarms : Send an alarm for correlation identification
     *
     * @param alarm  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "postAlarm",
        summary = "Send an alarm for correlation identification",
        tags = { "alarms" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = IOC.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/alarms",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<IOC> postAlarm(
        @Parameter(name = "Alarm", description = "", required = true) @Valid @RequestBody Alarm alarm
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"match\" : \"Data['email']['source'] == malicious.com\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"threat\" : { \"severity\" : \"high\", \"code\" : \"phishing\", \"technique\" : \"t1566.001\", \"tactic\" : \"ta0001\" } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}